Product Movement History is added

+++ src/main/java/org/bdlions/inventory/entity/EntityPOShowRoomProduct.java
@@ -52,10 +52,10 @@ public class EntityPOShowRoomProduct extends ClientResponse implements java.io.S
     private double unitPrice;
     
     @Column(name = "created_on", length = 11, columnDefinition = "int(11) unsigned DEFAULT 0")
-    private int createdOn;
+    private long createdOn;
 
     @Column(name = "modified_on", length = 11, columnDefinition = "int(11) unsigned DEFAULT 0")
-    private int modifiedOn;
+    private long modifiedOn;
 
     public int getId() {
         return id;
@@ -97,19 +97,19 @@ public class EntityPOShowRoomProduct extends ClientResponse implements java.io.S
         this.unitPrice = unitPrice;
     }
 
-    public int getCreatedOn() {
+    public long getCreatedOn() {
         return createdOn;
     }
 
-    public void setCreatedOn(int createdOn) {
+    public void setCreatedOn(long createdOn) {
         this.createdOn = createdOn;
     }
 
-    public int getModifiedOn() {
+    public long getModifiedOn() {
         return modifiedOn;
     }
 
-    public void setModifiedOn(int modifiedOn) {
+    public void setModifiedOn(long modifiedOn) {
         this.modifiedOn = modifiedOn;
     }
