+++ src/main/java/org/bdlions/inventory/entity/EntitySaleOrderProduct.java
@@ -62,10 +62,10 @@ public class EntitySaleOrderProduct extends ClientResponse implements java.io.Se
     private double vat;
     
     @Column(name = "created_on", length = 11, columnDefinition = "int(11) unsigned DEFAULT 0")
-    private int createdOn;
+    private long createdOn;
 
     @Column(name = "modified_on", length = 11, columnDefinition = "int(11) unsigned DEFAULT 0")
-    private int modifiedOn;
+    private long modifiedOn;
 
     public int getId() {
         return id;
@@ -115,19 +115,19 @@ public class EntitySaleOrderProduct extends ClientResponse implements java.io.Se
         this.discount = discount;
     }
 
-    public int getCreatedOn() {
+    public long getCreatedOn() {
         return createdOn;
     }
 
-    public void setCreatedOn(int createdOn) {
+    public void setCreatedOn(long createdOn) {
         this.createdOn = createdOn;
     }
 
-    public int getModifiedOn() {
+    public long getModifiedOn() {
         return modifiedOn;
     }
 
-    public void setModifiedOn(int modifiedOn) {
+    public void setModifiedOn(long modifiedOn) {
         this.modifiedOn = modifiedOn;
     }
